Create 03-11_1_PopulateMongoDatabase.txt
A script to populate your restaurants, add some menus, categories, and food items. 

I would suggest completely dropping your current "tables" and re-creating them with this script. 

I think we can continually upgrade this script with whatever the developer is working on and wants to share. Or maybe we make new files each time and keep the date naming syntax so we know which ones are more recent?